ENH,TST,DOC: Enhanced communications with better docs and testing.
Added pytest.ini and an environment._super_hydro_test.yml file for
creating a testing environment.  To run the tests just do:

    py.test

in the top level.  This only looks in the tests/ directory as
specified in pytest.ini.  Eventually it should also run tests in
super_hydro/.

--HG--
branch : client_server