ENH,TST,DOC: Enhanced communications with better docs and testing.
Added pytest.ini and an environment._super_hydro_test.yml file for
creating a testing environment.  To run the tests just do:

    py.test

in the top level.  This only looks in the tests/ directory as
specified in pytest.ini.  Eventually it should also run tests in
super_hydro/.

ENH: Run computation in a separate thread.

With this version of the code, the computation runs in a separate
thread allowing it to run independently of the client.  (Before, the
simulation would only run when the client requested an update.)

The client can now ask the simulation to pause, start, and quit.

server/:
- Computation runs in separate thread using Queues to communicate with
  the main communication server.
- Computation runs up to maximum specified fps.  (Should add some
  logging to test the actual achievable framerate.  Another
  enhancement could be to adjust the number of steps taken to match
  the desired fps if possible.)
- Two coordinate systems are used now: xy = data coordinates,
  pos = frame coordinates  - [0,1) as a fraction of the frame.

client/:
- Added some methods that get called when buttons are pressed to send
  the server commands like "Pause" and "Start".

- To Do: allow the client to send a "Quit" command.

--HG--
branch : client_server