BUG: Removed event-loop processing from withing sync() context.
This was causing a stack overflow because processing events within
this context would trigger updates which would again enter the sync()
context in a nested fashion.  This might adversely affect the performance
of the interface - need to test if the sleep() in sync allows
IPython to keep polling.  An alternative solution might be to put a
lock that prevents sync from being re-entrant.